+++ actix-framed/src/lib.rs
@@ -2,6 +2,7 @@ mod app;
 mod helpers;
 mod request;
 mod route;
+mod service;
 mod state;
 
 // re-export for convinience
@@ -10,4 +11,5 @@ pub use actix_http::{http, Error, HttpMessage, Response, ResponseError};
 pub use self::app::{FramedApp, FramedAppService};
 pub use self::request::FramedRequest;
 pub use self::route::FramedRoute;
+pub use self::service::{SendError, VerifyWebSockets};
 pub use self::state::State;